Move some misplaced tests under full package, add CDI_FULL category. Rework few other tests to work equally for Lite and Full.
