+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/alternative/selection/stereotype/Bar.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source
- * Copyright 2010, Red Hat, Inc., and individual contributors
+ * Copyright 2012, Red Hat, Inc., and individual contributors
  * by the @authors tag. See the copyright.txt in the distribution for a
  * full listing of individual contributors.
  *
@@ -9,21 +9,15 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,  
+ * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.implementation.disposal.method.definition.broken.interceptor;
 
-import jakarta.enterprise.context.Dependent;
-import jakarta.enterprise.inject.Produces;
+package org.jboss.cdi.tck.tests.full.alternative.selection.stereotype;
 
-@Dependent
-public class FooProducer {
-    @Produces
-    @ProducedString
-    public String produce() {
-        return "foo";
-    }
+@UnselectedStereotype
+public class Bar {
+
 }
