+++ impl/src/main/java/org/jboss/cdi/tck/tests/implementation/producer/method/broken/interceptor/ProducerMethodOnInterceptorTest.java
@@ -23,6 +23,7 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.container.test.api.ShouldThrowException;
 import org.jboss.cdi.tck.AbstractTest;
 import org.jboss.cdi.tck.shrinkwrap.WebArchiveBuilder;
+import org.jboss.shrinkwrap.api.BeanDiscoveryMode;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.jboss.shrinkwrap.impl.BeansXml;
 import org.jboss.test.audit.annotations.SpecAssertion;
@@ -37,7 +38,7 @@ public class ProducerMethodOnInterceptorTest extends AbstractTest {
     public static WebArchive createTestArchive() {
         return new WebArchiveBuilder()
                 .withTestClassPackage(ProducerMethodOnInterceptorTest.class)
-                .withBeansXml(new BeansXml().interceptors(SimpleInterceptor_Broken.class)).build();
+                .withBeansXml(new BeansXml(BeanDiscoveryMode.ANNOTATED)).build();
     }
 
     @Test
