--- impl/src/main/resources/org/jboss/cdi/tck/tests/definition/stereotype/inheritance/beans.xml
@@ -1,5 +1,0 @@
-<beans> 
-   <alternatives>
-      <stereotype>org.jboss.cdi.tck.tests.definition.stereotype.inheritance.NamedRequestPolicyStereotype</stereotype>
-   </alternatives>
-</beans>
