+++ impl/src/main/resources/org/jboss/cdi/tck/tests/full/alternative/broken/incorrect/name/beans.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<beans xmlns="https://jakarta.ee/xml/ns/jakartaee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="https://jakarta.ee/xml/ns/jakartaee
+         https://jakarta.ee/xml/ns/jakartaee/beans_3_0.xsd"
+       version="3.0" bean-discovery-mode="all">
+    <alternatives>
+        <class>org.jboss.cdi.tck.tests.full.alternative.broken.incorrect.name.NonExistingClass</class>
+    </alternatives>
+</beans>
