+++ impl/src/main/resources/org/jboss/cdi/tck/tests/full/alternative/broken/incorrect/name/stereotype/beans.xml
