+++ web/src/main/resources/org/jboss/cdi/tck/tests/full/interceptors/definition/interceptorNotListedInBeansXml/beans.xml
